Dash for R v0.6.3 (#213)
This release upgrades dash-renderer from v1.4.1 to v1.5.1, and is timed with the release of Dash for Python v1.13.4. This update also addresses a regression that prevented multiple loading states from displaying concurrently when a callback updates multiple outputs, as noted in [plotly/dash#1310](https://github.com/plotly/dash/pull/1310).